# server/openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RUN APPAREL Media & Asset Management API",
    "description": "Comprehensive API documentation for media asset management and folder organization endpoints. Handles images, 3D models (GLTF/GLB), and documents with advanced features including chunked uploads, batch operations, hierarchical folder management, event-driven cache invalidation, and analytics. Implements SAGA-inspired atomic deletion with compensating transactions, two-phase commit cache consistency, and stale-while-revalidate caching patterns for optimal performance and reliability.",
    "version": "2.0.0",
    "contact": {
      "name": "RUN APPAREL (PVT) LTD",
      "email": "admin@runapparel.lk"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Media & Asset Management API"
    }
  ],
...
  "paths": {
...
    "/cache/invalidation-time": {
      "get": {
        "tags": ["Cache Management"],
        "summary": "Get cache invalidation timestamp",
        "description": "Event-driven cache invalidation endpoint that returns the latest invalidation timestamp for a given pattern. Used by frontend polling to detect backend cache invalidations and trigger automatic UI updates. Part of the Phase 4 & 5 two-layer cache invalidation architecture (backend event bus + frontend polling).",
        "parameters": [
          {
            "name": "pattern",
            "in": "query",
            "required": true,
            "description": "Cache pattern to check (e.g., 'media:' for all media-related caches)",
            "schema": { "type": "string", "example": "media:" }
          }
        ],
        "responses": {
          "200": {
            "description": "Latest invalidation timestamp",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pattern": { "type": "string", "example": "media:" },
                    "timestamp": { "type": "integer", "example": 1729795200000, "description": "Unix timestamp in milliseconds of the latest invalidation event" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing required pattern parameter",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Resource not found" },
          "code": { "type": "string", "example": "NOT_FOUND" },
          "details": { "type": "object", "nullable": true }
        }
      }
    },
    "responses": {
...
      "ServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" },
            "example": {
              "success": false,
              "error": "Internal server error",
              "code": "INTERNAL_ERROR"
            }
          }
        }
      }
    }
  }
}